MAINT: Move ArrayVec implementation to module
